+++ src/Elements/Floor.cs
@@ -10,7 +10,7 @@ namespace Elements
     /// <summary>
     /// A floor is a horizontal element defined by a perimeter and one or several voids.
     /// </summary>
-    public class Floor : Element, IElementType<FloorType>, ISolid, IProfile
+    public class Floor : Element, IElementType<FloorType>, ISolid, IExtrude, IHasOpenings
     {
         /// <summary>
         /// The elevation from which the floor is extruded.
@@ -35,26 +35,22 @@ namespace Elements
         /// <summary>
         /// The openings in the floor.
         /// </summary>
-        public Opening[] Openings{get;}
+        public Opening[] Openings { get; }
 
         /// <summary>
-        /// Create a Floor.
+        /// The extrude direction of the floor.
         /// </summary>
-        /// <param name="profile">The profile of the floor.</param>
-        /// <param name="elementType">The floor type of the floor.</param>
-        /// <param name="elevation">The elevation of the top of the floor.</param>
-        /// <param name="transform">The floor's transform.</param>
-        /// <returns>A floor.</returns>
-        [JsonConstructor]
-        public Floor(Profile profile, FloorType elementType, double elevation = 0.0, Transform transform = null)
-        {
-            this.Profile = profile;
-            this.Elevation = elevation;
-            this.ElementType = elementType;
-            var thickness = elementType.Thickness();
-            this.Transform = transform != null ? transform : new Transform(new Vector3(0, 0, elevation - thickness));
-            this.Geometry = Solid.SweepFace(this.Profile.Perimeter, this.Profile.Voids, thickness);
-        }
+        public Vector3 ExtrudeDirection {get;}
+
+        /// <summary>
+        /// The extrude depth of the floor.
+        /// </summary>
+        public double ExtrudeDepth => this.Thickness();
+
+        /// <summary>
+        /// Extrude to both sides?
+        /// </summary>
+        public bool BothSides => false;
 
         /// <summary>
         /// Create a floor.
@@ -66,55 +62,50 @@ namespace Elements
         /// <param name="openings">An array of openings in the floor.</param>
         public Floor(Polygon profile, FloorType elementType, double elevation = 0.0, Transform transform = null, Opening[] openings = null)
         {
-            if (openings != null && openings.Length > 0)
-            {
-                var voids = new Polygon[openings.Length];
-                for (var i = 0; i < voids.Length; i++)
-                {
-                    var o = openings[i];
-                    voids[i] = o.Perimeter;
-                }
-                this.Profile = new Profile(profile, voids);
-            }
-            else
-            {
-                this.Profile = new Profile(profile);
-            }
-
+            this.Profile = new Profile(profile);
             this.Openings = openings;
             this.Elevation = elevation;
             this.ElementType = elementType;
             var thickness = elementType.Thickness();
-            this.Transform = transform != null ? transform : new Transform(new Vector3(0, 0, elevation - thickness));
-            this.Geometry = Solid.SweepFace(this.Profile.Perimeter, this.Profile.Voids, thickness);
+            this.Transform = transform != null ? transform : new Transform(new Vector3(0, 0, elevation));
+            this.ExtrudeDirection = Vector3.ZAxis;
         }
 
         /// <summary>
         /// Create a floor.
         /// </summary>
         /// <param name="profile">The profile of the floor.</param>
-        /// <param name="start">A tranforms used to pre-transform the profile and direction vector before sweeping the geometry.</param>
+        /// <param name="start">A tranform used to pre-transform the profile and direction vector before sweeping the geometry.</param>
         /// <param name="direction">The direction of the floor's sweep.</param>
         /// <param name="elementType">The floor type of the floor.</param>
         /// <param name="elevation">The elevation of the floor.</param>
         /// <param name="transform">The floor's transform. If set, this will override the elevation.</param>
         public Floor(Profile profile, Transform start, Vector3 direction, FloorType elementType, double elevation = 0.0, Transform transform = null)
         {
-            this.Profile = profile;
+            // this.Profile = new Profile(profile);
             this.Elevation = elevation;
             this.ElementType = elementType;
             this.Transform = transform != null ? transform : new Transform(new Vector3(0, 0, elevation));
-            var outer = start.OfPolygon(this.Profile.Perimeter);
-            var inner = this.Profile.Voids != null ? start.OfPolygons(this.Profile.Voids) : null;
-            this.Geometry = Solid.SweepFace(outer, inner, start.OfVector(direction), this.Thickness());
+            this.Profile = start.OfProfile(profile);
+            this.ExtrudeDirection = start.OfVector(direction);
         }
-        
+
+        [JsonConstructor]
+        internal Floor(Profile profile, FloorType elementType, double elevation = 0.0, Transform transform = null, Opening[] openings = null)
+        {
+            this.Profile = profile;
+            this.Openings = openings;
+            this.Elevation = elevation;
+            this.ElementType = elementType;
+            var thickness = elementType.Thickness();
+            this.Transform = transform != null ? transform : new Transform(new Vector3(0, 0, elevation));
+            this.ExtrudeDirection = Vector3.ZAxis;
+        }
+
         /// <summary>
         /// The area of the floor.
-        /// Overlapping openings and openings which are outside of the floor's perimeter,
-        /// will result in incorrect area results.
         /// </summary>
-        /// <returns>The area of the floor.</returns>
+        /// <returns>The area of the floor, not including the area of openings.</returns>
         public double Area()
         {
             return this.Profile.Area();