+++ src/Elements/Geometry/Solids/Solid.cs
@@ -1,6 +1,7 @@
 #pragma warning disable 1591
 
 using System.Collections.Generic;
+using System.Linq;
 using System.Runtime.CompilerServices;
 using System.Text;
 using Elements.Geometry.Interfaces;
@@ -69,26 +70,26 @@ namespace Elements.Geometry.Solids
         /// <summary>
         /// Construct a solid by sweeping a face.
         /// </summary>
-        /// <param name="outerLoop">The perimeter of the face to sweep.</param>
-        /// <param name="innerLoops">The holes of the face to sweep.</param>
+        /// <param name="perimeter">The perimeter of the face to sweep.</param>
+        /// <param name="holes">The holes of the face to sweep.</param>
         /// <param name="distance">The distance to sweep.</param>
         /// <param name="bothSides">Should the sweep start offset by direction distance/2? </param>
         /// <returns>A solid.</returns>
-        public static Solid SweepFace(Polygon outerLoop, Polygon[] innerLoops, double distance, bool bothSides = false) 
+        public static Solid SweepFace(Polygon perimeter, Polygon[] holes, double distance, bool bothSides = false) 
         {
-            return Solid.SweepFace(outerLoop, innerLoops, Vector3.ZAxis, distance, bothSides);
+            return Solid.SweepFace(perimeter, holes, Vector3.ZAxis, distance, bothSides);
         }
 
         /// <summary>
         /// Construct a solid by sweeping a face along a curve.
         /// </summary>
-        /// <param name="outer">The perimeter of the face to sweep.</param>
-        /// <param name="inner">The holes of the face to sweep.</param>
+        /// <param name="perimeter">The perimeter of the face to sweep.</param>
+        /// <param name="holes">The holes of the face to sweep.</param>
         /// <param name="curve">The curve along which to sweep.</param>
         /// <param name="startSetback">The setback of the sweep from the start of the curve.</param>
         /// <param name="endSetback">The setback of the sweep from the end of the curve.</param>
         /// <returns>A solid.</returns>
-        public static Solid SweepFaceAlongCurve(Polygon outer, Polygon[] inner, ICurve curve,  double startSetback = 0, double endSetback = 0)
+        public static Solid SweepFaceAlongCurve(Polygon perimeter, Polygon[] holes, ICurve curve,  double startSetback = 0, double endSetback = 0)
         {
             var solid = new Solid();
 
@@ -103,7 +104,7 @@ namespace Elements.Geometry.Solids
                 for (var i = 0; i < transforms.Length; i++)
                 {
                     var next = i == transforms.Length - 1 ? transforms[0] : transforms[i + 1];
-                    solid.SweepPolygonBetweenPlanes(outer, transforms[i].XY(), next.XY());
+                    solid.SweepPolygonBetweenPlanes(perimeter, transforms[i].XY(), next.XY());
                 }
             }
             else
@@ -112,14 +113,14 @@ namespace Elements.Geometry.Solids
                 Face cap = null;
                 Edge[][] openEdges;
 
-                if(inner != null)
+                if(holes != null)
                 {
-                    cap = solid.AddFace(transforms[0].OfPolygon(outer), transforms[0].OfPolygons(inner));
-                    openEdges = new Edge[1 + inner.Length][];
+                    cap = solid.AddFace(transforms[0].OfPolygon(perimeter), transforms[0].OfPolygons(holes));
+                    openEdges = new Edge[1 + holes.Length][];
                 }
                 else
                 {
-                    cap = solid.AddFace(transforms[0].OfPolygon(outer));
+                    cap = solid.AddFace(transforms[0].OfPolygon(perimeter));
                     openEdges = new Edge[1][];
                 }
 
@@ -128,7 +129,7 @@ namespace Elements.Geometry.Solids
                 openEdge = solid.SweepEdges(transforms, openEdge);
                 openEdges[0] = openEdge;
 
-                if(inner != null)
+                if(holes != null)
                 {
                     for(var i=0; i<cap.Inner.Length; i++)
                     {
@@ -149,46 +150,56 @@ namespace Elements.Geometry.Solids
         /// <summary>
         /// Construct a solid by sweeping a face in a direction.
         /// </summary>
-        /// <param name="outerLoop">The perimeter of the face to sweep.</param>
-        /// <param name="innerLoops">The holes of the face to sweep.</param>
+        /// <param name="perimeter">The perimeter of the face to sweep.</param>
+        /// <param name="holes">The holes of the face to sweep.</param>
         /// <param name="direction">The direction in which to sweep.</param>
         /// <param name="distance">The distance to sweep.</param>
         /// <param name="bothSides">Should the sweep start offset by direction distance/2? </param>
         /// <returns>A solid.</returns>
-        public static Solid SweepFace(Polygon outerLoop, Polygon[] innerLoops, Vector3 direction, double distance, bool bothSides = false)
+        public static Solid SweepFace(Polygon perimeter, Polygon[] holes, Vector3 direction, double distance, bool bothSides = false)
         {
+            // We do a difference of the polygons
+            // to get the clipped shape. This will fail in interesting
+            // ways if the clip creates two islands.
+            if(holes != null)
+            {
+                var newPerimeter = perimeter.Difference(holes);
+                perimeter = newPerimeter[0];
+                holes = newPerimeter.Skip(1).Take(newPerimeter.Count - 1).ToArray();
+            }
+            
             var solid = new Solid();
             Face fStart = null;
             if(bothSides)
             {
                 var t = new Transform(direction.Negated()* (distance/2));
-                if(innerLoops != null)
+                if(holes != null)
                 {
-                    fStart = solid.AddFace(t.OfPolygon(outerLoop.Reversed()), t.OfPolygons(innerLoops.Reversed()));
+                    fStart = solid.AddFace(t.OfPolygon(perimeter.Reversed()), t.OfPolygons(holes.Reversed()));
                 }
                 else
                 {
-                    fStart = solid.AddFace(t.OfPolygon(outerLoop.Reversed()));
+                    fStart = solid.AddFace(t.OfPolygon(perimeter.Reversed()));
                 }
             }
             else
             {
-                if(innerLoops != null)
+                if(holes != null)
                 {
-                    fStart = solid.AddFace(outerLoop.Reversed(), innerLoops.Reversed());
+                    fStart = solid.AddFace(perimeter.Reversed(), holes.Reversed());
                 }
                 else
                 {
-                    fStart = solid.AddFace(outerLoop.Reversed());
+                    fStart = solid.AddFace(perimeter.Reversed());
                 }
             }
 
             var fEndOuter = solid.SweepLoop(fStart.Outer, direction, distance);
 
-            if(innerLoops != null)
+            if(holes != null)
             {
-                var fEndInner = new Loop[innerLoops.Length];
-                for(var i=0; i<innerLoops.Length; i++)
+                var fEndInner = new Loop[holes.Length];
+                for(var i=0; i<holes.Length; i++)
                 {
                     fEndInner[i] = solid.SweepLoop(fStart.Inner[i], direction, distance);
                 }