+++ src/Elements/Serialization/IFC/IFCExtensions.cs
@@ -20,18 +20,45 @@ namespace Elements.Serialization.IFC
         /// Load a model from IFC.
         /// </summary>
         /// <param name="path">The path to an IFC STEP file.</param>
+        /// <param name="idsToConvert">An array of element ids to convert.</param>
         /// <returns>A model.</returns>
-        internal static Model FromIFC(string path)
+        internal static Model FromIFC(string path, IList<string> idsToConvert = null)
         {
             List<STEPError> errors;
             var ifcModel = new Document(path, out errors);
-            var ifcSlabs = ifcModel.AllInstancesOfType<IfcSlab>();
-            var ifcSpaces = ifcModel.AllInstancesOfType<IfcSpace>();
-            var ifcWalls = ifcModel.AllInstancesOfType<IfcWallStandardCase>();
-            var ifcBeams = ifcModel.AllInstancesOfType<IfcBeam>();
-            var ifcColumns = ifcModel.AllInstancesOfType<IfcColumn>();
-            var ifcVoids = ifcModel.AllInstancesOfType<IfcRelVoidsElement>();
-            var ifcMaterials = ifcModel.AllInstancesOfType<IfcRelAssociatesMaterial>();
+            foreach(var error in errors)
+            {
+                Console.WriteLine("***IFC ERROR***" + error.Message);
+            }
+
+            IEnumerable<IfcSlab> ifcSlabs = null;
+            IEnumerable<IfcSpace> ifcSpaces = null;
+            IEnumerable<IfcWallStandardCase> ifcWalls = null;
+            IEnumerable<IfcBeam> ifcBeams = null;
+            IEnumerable<IfcColumn> ifcColumns = null;
+            IEnumerable<IfcRelVoidsElement> ifcVoids = null;
+            IEnumerable<IfcRelAssociatesMaterial> ifcMaterials = null;
+
+            if(idsToConvert != null && idsToConvert.Count > 0)
+            {
+                ifcSlabs = ifcModel.AllInstancesOfType<IfcSlab>().Where(i=>idsToConvert.Contains(i.GlobalId));
+                ifcSpaces = ifcModel.AllInstancesOfType<IfcSpace>().Where(i=>idsToConvert.Contains(i.GlobalId));
+                ifcWalls = ifcModel.AllInstancesOfType<IfcWallStandardCase>().Where(i=>idsToConvert.Contains(i.GlobalId));
+                ifcBeams = ifcModel.AllInstancesOfType<IfcBeam>().Where(i=>idsToConvert.Contains(i.GlobalId));
+                ifcColumns = ifcModel.AllInstancesOfType<IfcColumn>().Where(i=>idsToConvert.Contains(i.GlobalId));
+                ifcVoids = ifcModel.AllInstancesOfType<IfcRelVoidsElement>().Where(i=>idsToConvert.Contains(i.GlobalId));
+                ifcMaterials = ifcModel.AllInstancesOfType<IfcRelAssociatesMaterial>().Where(i=>idsToConvert.Contains(i.GlobalId));
+            }
+            else
+            {
+                ifcSlabs = ifcModel.AllInstancesOfType<IfcSlab>();
+                ifcSpaces = ifcModel.AllInstancesOfType<IfcSpace>();
+                ifcWalls = ifcModel.AllInstancesOfType<IfcWallStandardCase>();
+                ifcBeams = ifcModel.AllInstancesOfType<IfcBeam>();
+                ifcColumns = ifcModel.AllInstancesOfType<IfcColumn>();
+                ifcVoids = ifcModel.AllInstancesOfType<IfcRelVoidsElement>();
+                ifcMaterials = ifcModel.AllInstancesOfType<IfcRelAssociatesMaterial>();
+            }
 
             var openings = new List<Opening>();
             foreach (var v in ifcVoids)
@@ -58,10 +85,10 @@ namespace Elements.Serialization.IFC
             model.AddElements(walls);
             model.AddElements(beams);
             model.AddElements(columns);
-            if (openings.Any())
-            {
-                model.AddElements(openings);
-            }
+            // if (openings.Any())
+            // {
+            //     model.AddElements(openings);
+            // }
 
             return model;
         }
@@ -112,9 +139,9 @@ namespace Elements.Serialization.IFC
             {
                 try
                 {
-                    if(e is Wall)
+                    if(e is StandardWall)
                     {
-                        var w = (Wall)e;
+                        var w = (StandardWall)e;
                         var ifcWall = w.ToIfcWallStandardCase(context, ifc);
                         products.Add(ifcWall);
                     }
@@ -133,8 +160,9 @@ namespace Elements.Serialization.IFC
                         products.Add(ifcSlab);
                     }
                 }
-                catch
+                catch(Exception ex)
                 {
+                    Console.WriteLine("There was an error writing to IFC: " + ex.Message);
                     continue;
                 }
             }
@@ -281,21 +309,22 @@ namespace Elements.Serialization.IFC
         /// <param name="context"></param>
         /// <param name="doc"></param>
         /// <returns></returns>
-        private static IfcWallStandardCase ToIfcWallStandardCase(this Wall wall, IfcRepresentationContext context, Document doc)
+        private static IfcWallStandardCase ToIfcWallStandardCase(this StandardWall wall, IfcRepresentationContext context, Document doc)
         {
             var sweptArea = wall.CenterLine.Thicken(wall.Thickness()).ToIfcArbitraryClosedProfileDef(doc);
             var extrudeDirection = Vector3.ZAxis.ToIfcDirection();
 
             // We don't use the Wall's transform for positioning, because
             // our walls have a transform that lays the wall "flat". Just
-            // use a identity transform.
+            // use a identity transform or a transform that includes
+            // the elevation.
             var position = new Transform().ToIfcAxis2Placement3D(doc);
             var repItem = new IfcExtrudedAreaSolid(sweptArea, position, 
                 extrudeDirection, new IfcPositiveLengthMeasure(wall.Height));
             var rep = new IfcShapeRepresentation(context, "Body", "SweptSolid", new List<IfcRepresentationItem>{repItem});
             var productRep = new IfcProductDefinitionShape(new List<IfcRepresentation>{rep});
             var id = IfcGuid.ToIfcGuid(Guid.NewGuid());
-            var localPlacement = new Transform().ToIfcLocalPlacement(doc);
+            var localPlacement = new Transform(0,0,wall.CenterLine.Start.Z).ToIfcLocalPlacement(doc);
             var ifcWall = new IfcWallStandardCase(new IfcGloballyUniqueId(id), 
                 null, wall.Name, null, null, localPlacement, productRep, null);
 
@@ -329,6 +358,7 @@ namespace Elements.Serialization.IFC
             var repItem = new IfcExtrudedAreaSolid(sweptArea, position, 
                 extrudeDirection, new IfcPositiveLengthMeasure(beam.Curve.Length()));
             var localPlacement = beam.Curve.TransformAt(0.0).ToIfcLocalPlacement(doc);
+            // var placement = beam.Transform.ToIfcAxis2Placement3D(doc);
             var rep = new IfcShapeRepresentation(context, "Body", "SweptSolid", new List<IfcRepresentationItem>{repItem});
             var productRep = new IfcProductDefinitionShape(new List<IfcRepresentation>{rep});
             var ifcBeam = new IfcBeam(IfcGuid.ToIfcGuid(Guid.NewGuid()), null, null, null, null, localPlacement, productRep, null);
@@ -459,7 +489,7 @@ namespace Elements.Serialization.IFC
                 // throw new Exception("No IfcExtrudedAreaSolid could be found in the provided IfcSlab.");
             }
 
-            var floorType = new FloorType($"{Guid.NewGuid().ToString()}_floor_type", (IfcLengthMeasure)solid.Depth);
+            var floorType = new FloorType($"{Guid.NewGuid().ToString()}_floor_type", new List<MaterialLayer>{new MaterialLayer(new Material("slab",Colors.Green), (IfcLengthMeasure)solid.Depth)});
             var outline = (Polygon)solid.SweptArea.ToICurve();
             var solidTransform = solid.Position.ToTransform();
             var floor = new Floor(new Profile(outline), solidTransform, solid.ExtrudedDirection.ToVector3(), 
@@ -538,6 +568,24 @@ namespace Elements.Serialization.IFC
 
             // An extruded face solid.
             var solid = wall.RepresentationsOfType<IfcExtrudedAreaSolid>().FirstOrDefault();
+            if(solid == null)
+            {
+                // It's possible that the rep is a boolean.
+                var boolean = wall.RepresentationsOfType<IfcBooleanClippingResult>().FirstOrDefault();
+                if(boolean != null)
+                {
+                    solid = boolean.FirstOperand.Choice as IfcExtrudedAreaSolid;
+                    if(solid == null)
+                    {
+                        solid = boolean.SecondOperand.Choice as IfcExtrudedAreaSolid; 
+                    }
+                }
+
+                // if(solid == null)
+                // {
+                //     throw new Exception("No usable solid was found when converting an IfcWallStandardCase to a Wall.");
+                // }
+            }
             
             // A centerline wall with material layers.
             // var axis = (Polyline)wall.RepresentationsOfType<IfcPolyline>().FirstOrDefault().ToICurve(false);
@@ -555,7 +603,7 @@ namespace Elements.Serialization.IFC
                 if(c is Polygon)
                 {
                     transform.Concatenate(solid.Position.ToTransform());
-                    var result = new Wall(new Profile((Polygon)c), wallType, (IfcLengthMeasure)solid.Depth, transform);
+                    var result = new Wall((Polygon)c, wallType, (IfcLengthMeasure)solid.Depth, transform);
 
                     result.Name = wall.Name;
                     return result;
@@ -637,7 +685,6 @@ namespace Elements.Serialization.IFC
                 solidTransform.Concatenate(openingTransform);
                 var profile = (Polygon)s.SweptArea.ToICurve();
                 var newOpening = new Opening(profile, (IfcLengthMeasure)s.Depth, solidTransform);
-                // var newOpening = new Opening(profile);
                 if(opening.Name != null)
                 {
                     newOpening.Name = opening.Name;