+++ test/FloorTests.cs
@@ -12,28 +12,22 @@ namespace Elements.Tests
         public void Floor()
         {
             this.Name = "Floor";
-            var p = Polygon.Rectangle(10, 10);
+            var p = Polygon.L(10, 20, 5);
             var floorType = new FloorType("test", new List<MaterialLayer>{new MaterialLayer(new Material("green", Colors.Green, 0.0f,0.0f),0.1)});
             var openings = new Opening[]{
                 new Opening(1, 1, 1, 1),
-                new Opening(-2, 3, 3, 1),
-                new Opening(2, -2, 1, 4),
+                new Opening(3, 3, 1, 3),
             };
             var floor1 = new Floor(p, floorType, 0.5, null, openings);
-            var floor2 = new Floor(p, floorType, 3.5, null, openings);
 
             var model = new Model();
-            
-            var s = new Space(Polygon.Rectangle(1,1), 1, 0.5, BuiltInMaterials.Mass);
-            this.Model.AddElement(s);
 
-            Assert.Equal(3, floor1.Openings.Length);
+            Assert.Equal(2, floor1.Openings.Length);
             Assert.Equal(0.5, floor1.Elevation);
             Assert.Equal(0.1, floor1.ElementType.Thickness());
-            Assert.Equal(0.4, floor1.Transform.Origin.Z);
+            Assert.Equal(0.5, floor1.Transform.Origin.Z);
 
             this.Model.AddElement(floor1);
-            this.Model.AddElement(floor2);
         }
 
         [Fact]
@@ -48,12 +42,13 @@ namespace Elements.Tests
         public void Area()
         {
             // A floor with two holes punched in it.
-            var p1 = Polygon.Rectangle(1, 1, new Vector3(1,1,0));
-            var p2 = Polygon.Rectangle(1, 1, new Vector3(3,3,0));
-            var profile = new Profile(Polygon.Rectangle(10, 10), new[]{p1,p2});
+            var p1 = Polygon.Rectangle(1, 1);
+            var p2 = Polygon.Rectangle(1, 1);
+            var o1 = new Opening(p1, 1, 1);
+            var o2 = new Opening(p2, 3, 3);
             var floorType = new FloorType("test", 0.2);
-            var floor = new Floor(profile, floorType, 0.0);
-            Assert.Equal(100.0-2.0, floor.Area());
+            var floor = new Floor(Polygon.Rectangle(10, 10), floorType, 0.0, null, new []{o1,o2});
+            Assert.Equal(100.0, floor.Area());
         }
     }
 }