+++ test/ModelTest.cs
@@ -30,6 +30,12 @@ namespace Elements.Tests
             set => _name = value;
         }
 
+        public bool GenerateIfc{get;set;}
+
+        public bool GenerateGlb{get;set;}
+
+        public bool GenerateJson{get;set;}
+
         internal static Line TestLine = new Line(Vector3.Origin, new Vector3(5,5,5));
         internal static Arc TestArc = new Arc(new Plane(Vector3.Origin, Vector3.ZAxis), 2.0, 0.0, 90.0);
         internal static Polyline TestPolyline = new Polyline(new []{new Vector3(0,0), new Vector3(0,2), new Vector3(0,3,1)});
@@ -38,6 +44,9 @@ namespace Elements.Tests
         public ModelTest()
         {
             this._model = new Model();
+            this.GenerateGlb = true;
+            this.GenerateIfc = true;
+            this.GenerateJson = true;
         }
 
         public void Dispose()
@@ -51,25 +60,34 @@ namespace Elements.Tests
                 var sw = new Stopwatch();
                 sw.Start();
 
-                // Write the model as a glb
-                var modelPath = $"models/{this._name}.glb";
-                this._model.ToGlTF($"models/{this._name}.glb");
-                sw.Stop();
-                Console.WriteLine($"Saved {this._name} to glb: {modelPath}.({sw.Elapsed.TotalMilliseconds}ms)");
+                if(this.GenerateGlb)
+                {
+                    // Write the model as a glb
+                    var modelPath = $"models/{this._name}.glb";
+                    this._model.ToGlTF($"models/{this._name}.glb");
+                    sw.Stop();
+                    Console.WriteLine($"Saved {this._name} to glb: {modelPath}.({sw.Elapsed.TotalMilliseconds}ms)");
+                }
 
-                // Write the model as json
-                var jsonPath = $"models/{this._name}.json";
-                Console.WriteLine($"Serializing {this._name} to JSON: {jsonPath}");
-                File.WriteAllText(jsonPath, this._model.ToJson());
+                if(this.GenerateJson)
+                {
+                    // Write the model as json
+                    var jsonPath = $"models/{this._name}.json";
+                    Console.WriteLine($"Serializing {this._name} to JSON: {jsonPath}");
+                    File.WriteAllText(jsonPath, this._model.ToJson());
 
-                // Try deserializing JSON
-                Console.WriteLine($"Deserializing {this._name} from JSON.");
-                var newModel = Model.FromJson(File.ReadAllText(jsonPath));
+                    // Try deserializing JSON
+                    Console.WriteLine($"Deserializing {this._name} from JSON.");
+                    var newModel = Model.FromJson(File.ReadAllText(jsonPath));
+                }
 
-                // Write the model as an IFC
-                var ifcPath = $"models/{this._name}.ifc";
-                Console.WriteLine($"Serializing {this._name} to IFC: {ifcPath}");
-                this._model.ToIFC(ifcPath);
+                if(this.GenerateIfc)
+                {
+                    // Write the model as an IFC
+                    var ifcPath = $"models/{this._name}.ifc";
+                    Console.WriteLine($"Serializing {this._name} to IFC: {ifcPath}");
+                    this._model.ToIFC(ifcPath);
+                }
             }
         }
     }