+++ test/WallTests.cs
@@ -11,32 +11,20 @@ namespace Elements.Tests
         [Fact]
         public void Wall()
         {
-            this.Name = "WallLinear";
+            this.Name = "Wall";
             var testWallType = new WallType("test", new List<MaterialLayer>{new MaterialLayer(new Material("blue", Colors.Blue, 0.0f, 0.0f), 0.1)});
 
-            var ngon = Polygon.Ngon(7, 15.0);
+            var l = new Line(new Vector3(0,0,0), new Vector3(10,10,0));
             var openings = new Opening[]{
                 new Opening(1.0, 2.0, 1.0, 1.0),
                 new Opening(3.0, 1.0, 1.0, 2.0),
-                // new Opening(7.0, 1.25, 3.0, 2.5),
-                new Opening(Polygon.Ngon(5, 2.0), 8,2)
+                new Opening(Polygon.Ngon(3, 2.0), 8,2)
             };
 
             var frameProfile = new Profile(Polygon.Rectangle(0.075, 0.01));
-            foreach(var l in ngon.Segments())
-            {
-                var w = new Wall(l, testWallType, 3.0, openings);
-                this.Model.AddElement(w);
 
-                // Draw some frames and panels in the openings.
-                foreach(var o in w.Openings)
-                {
-                    var f = new Frame(o.Perimeter, frameProfile, 0.0375, BuiltInMaterials.Black, w.Transform);
-                    this.Model.AddElement(f);
-                    var p = new Panel(o.Perimeter, BuiltInMaterials.Glass, w.Transform);
-                    this.Model.AddElement(p);
-                }
-            }
+            var w = new StandardWall(l, testWallType, 3.0, openings);
+            this.Model.AddElement(w);
         }
 
         [Fact]
@@ -46,7 +34,7 @@ namespace Elements.Tests
             var b = new Vector3(0.0, 5.0);
             var line = new Line(a,b);
             var testWallType = new WallType("test", 0.1);
-            Assert.Throws<ArgumentOutOfRangeException>(()=>new Wall(line, testWallType, 0.0));
+            Assert.Throws<ArgumentOutOfRangeException>(()=>new StandardWall(line, testWallType, 0.0));
         }
 
         [Fact]
@@ -65,7 +53,7 @@ namespace Elements.Tests
             var b = new Vector3(0.0, 5.0, 5.0);
             var line = new Line(a,b);
             var testWallType = new WallType("test", 0.1);
-            Assert.Throws<ArgumentException>(()=>new Wall(line, testWallType, 5.0));
+            Assert.Throws<ArgumentException>(()=>new StandardWall(line, testWallType, 5.0));
         }
 
         [Fact]
@@ -75,8 +63,8 @@ namespace Elements.Tests
             var b = new Vector3(0.0, 5.0);
             var line = new Line(a,b);
             var testWallType = new WallType("test", 0.1);
-            var wall = new Wall(line, testWallType, 4.0);
-            Assert.Null(wall.Profile.Voids);
+            var wall = new StandardWall(line, testWallType, 4.0);
+            Assert.Null(wall.Openings);
         }
     }
 }